ai/tvds-ad/requirements.txt
numpy~=1.23.3
timm~=0.6.11
pillow~=9.2.0
sklearn~=0.0.post1
scikit-learn~=1.1.3
torchvision~=0.14.0